Add database project for RightsU, RoP, Avail

+++ RightsU_DB/dbo/Stored Procedures/USP_BMS_WBS_LIst.sql
@@ -1,5 +1,4 @@
-﻿
-CREATE PROCEDURE [dbo].[USP_BMS_WBS_LIst]
+﻿CREATE PROCEDURE [dbo].[USP_BMS_WBS_LIst]
 (
 	@Is_Process VARCHAR(5),
 	@WBS_Code NVARCHAR(100),
@@ -13,45 +12,51 @@ AS
 -- =============================================
 
 BEGIN	
-	SET NOCOUNT ON;
+	Declare @Loglevel  int;
 
-    SELECT [BMS_WBS_Code]
-      ,[SAP_WBS_Code]
-      ,[WBS_Code]
-      ,[WBS_Description]
-      ,[Short_ID]
-	  ,CASE
-		 WHEN ISNULL([Short_ID] ,'') <> '' THEN [WBS_Description] + '-' + [Short_ID] + '-' + [WBS_Code] 
-			ELSE  [WBS_Description] +  '-' + [WBS_Code] 
-	  END
-		AS  WBS_Description_Short_ID
-      ,[Is_Archive]
-      ,[Status]
-      ,ISNULL([Error_Details],'') AS Error_Details
-      ,[Is_Process]
-      ,[File_Code]
-	  ,[BMS_Key]
-	  ,ISNULL([Response_Type],'') AS Response_Type
-	  ,ISNULL([Response_Status],'') AS Response_Status
-	  INTO #Temp
-  FROM [dbo].[BMS_WBS]
-  WHERE [Is_Process] = @Is_Process
-  --AND ISNULL([BMS_Key],0) > 0
+	select @Loglevel = Parameter_Value from System_Parameter_New where Parameter_Name='loglevel'
 
-	UPDATE #Temp SET WBS_Description_Short_ID = 
-	CASE 
-		WHEN LEN(WBS_Description_Short_ID) > 80
-			THEN  -- SUBSTRING(WBS_Description ,1, (LEN(WBS_Description) - (LEN(WBS_Description_Short_ID) - 80)))  + '-'			 
-			CASE
-				 WHEN ISNULL([Short_ID] ,'') <> '' 
-					THEN SUBSTRING(WBS_Description ,1, (LEN(WBS_Description) - (LEN(WBS_Description_Short_ID) - 80))) + '-' + [Short_ID] + '-' + [WBS_Code] 
-					ELSE  SUBSTRING(WBS_Description ,1, (LEN(WBS_Description) - (LEN(WBS_Description_Short_ID) - 80))) +  '-' + [WBS_Code] 
-			END
-        ELSE WBS_Description_Short_ID
-	END  
+	if(@Loglevel  < 2)Exec [USPLogSQLSteps] '[USP_BMS_WBS_LIst]', 'Step 1', 0, 'Started Procedure', 0, ''
+		SET NOCOUNT ON;
 
-	SELECT * FROM #Temp
+		SELECT [BMS_WBS_Code]
+		  ,[SAP_WBS_Code]
+		  ,[WBS_Code]
+		  ,[WBS_Description]
+		  ,[Short_ID]
+		  ,CASE
+			 WHEN ISNULL([Short_ID] ,'') <> '' THEN [WBS_Description] + '-' + [Short_ID] + '-' + [WBS_Code] 
+				ELSE  [WBS_Description] +  '-' + [WBS_Code] 
+		  END
+			AS  WBS_Description_Short_ID
+		  ,[Is_Archive]
+		  ,[Status]
+		  ,ISNULL([Error_Details],'') AS Error_Details
+		  ,[Is_Process]
+		  ,[File_Code]
+		  ,[BMS_Key]
+		  ,ISNULL([Response_Type],'') AS Response_Type
+		  ,ISNULL([Response_Status],'') AS Response_Status
+		  INTO #Temp
+	  FROM [dbo].[BMS_WBS] (NOLOCK)
+	  WHERE [Is_Process] = @Is_Process
+	  --AND ISNULL([BMS_Key],0) > 0
 
-	IF OBJECT_ID('tempdb..#Temp') IS NOT NULL DROP TABLE #Temp
+		UPDATE #Temp SET WBS_Description_Short_ID = 
+		CASE 
+			WHEN LEN(WBS_Description_Short_ID) > 80
+				THEN  -- SUBSTRING(WBS_Description ,1, (LEN(WBS_Description) - (LEN(WBS_Description_Short_ID) - 80)))  + '-'			 
+				CASE
+					 WHEN ISNULL([Short_ID] ,'') <> '' 
+						THEN SUBSTRING(WBS_Description ,1, (LEN(WBS_Description) - (LEN(WBS_Description_Short_ID) - 80))) + '-' + [Short_ID] + '-' + [WBS_Code] 
+						ELSE  SUBSTRING(WBS_Description ,1, (LEN(WBS_Description) - (LEN(WBS_Description_Short_ID) - 80))) +  '-' + [WBS_Code] 
+				END
+			ELSE WBS_Description_Short_ID
+		END  
 
+		SELECT * FROM #Temp
+
+		IF OBJECT_ID('tempdb..#Temp') IS NOT NULL DROP TABLE #Temp
+
+	if(@Loglevel  < 2)Exec [USPLogSQLSteps] '[USP_BMS_WBS_LIst]', 'Step 2', 0, 'Procedure Excution Completed', 0, ''
 END
